+++ src/ultimate_rvc/rvc/train/extract/extract.py
@@ -24,6 +24,7 @@ from ultimate_rvc.common import RVC_MODELS_DIR, lazy_import
 from ultimate_rvc.rvc.configs.config import Config
 from ultimate_rvc.rvc.lib.predictors.RMVPE import RMVPE0Predictor
 from ultimate_rvc.rvc.lib.utils import load_audio, load_embedding
+from ultimate_rvc.rvc.train.utils import remove_sox_libmso6_from_ld_preload
 
 if TYPE_CHECKING:
     import static_sox.run as static_sox_run
@@ -136,32 +137,6 @@ class FeatureInput:
                     pbar.update(1)
 
 
-def remove_from_ld_preload(prefix: str) -> None:
-    """
-    Remove entries from the LD_PRELOAD environment variable that start
-    with the given prefix.
-
-    Parameters
-    ----------
-    prefix : str
-        The prefix to match entries in LD_PRELOAD.
-
-    """
-    # Get the current LD_PRELOAD value
-    ld_preload = os.environ.get("LD_PRELOAD", "")
-
-    # Split the LD_PRELOAD into a list of entries
-    preload_entries = ld_preload.split(os.pathsep)
-
-    # Remove the entries that start with the given prefix
-    preload_entries = [
-        entry for entry in preload_entries if not entry.startswith(prefix)
-    ]
-
-    # Join the list back into a string and update LD_PRELOAD
-    os.environ["LD_PRELOAD"] = os.pathsep.join(preload_entries)
-
-
 def run_pitch_extraction(
     files: list[list[str]],
     devices: list[str],
@@ -179,13 +154,7 @@ def run_pitch_extraction(
     start_time = time.time()
     fe = FeatureInput()
 
-    # NOTE On ubuntu 24.04 the static_sox module does not work with
-    # multiprocessing using the spawn method due to a "version
-    # `GLIBC_2.38' not found" error. This is a workaround, which removes
-    # the path to the libm.so.6 library from the LD_PRELOAD environment
-    # variable.
-    sox_exe = static_sox_run.get_or_fetch_platform_executables_else_raise()
-    remove_from_ld_preload(os.path.join(os.path.dirname(sox_exe), "libm.so.6"))
+    remove_sox_libmso6_from_ld_preload()
 
     with concurrent.futures.ProcessPoolExecutor(max_workers=len(devices)) as executor:
         tasks = [
