+++ src/ultimate_rvc/rvc/train/utils.py
@@ -1,3 +1,5 @@
+from typing import TYPE_CHECKING
+
 import glob
 import logging
 import os
@@ -10,9 +12,58 @@ import torch
 
 import soundfile as sf
 
+from ultimate_rvc.common import lazy_import
+
+if TYPE_CHECKING:
+    import static_sox.run as static_sox_run
+else:
+    static_sox_run = lazy_import("static_sox.run")
+
 logger = logging.getLogger(__name__)
 
 MATPLOTLIB_FLAG = False
+
+
+def remove_from_ld_preload(prefix: str) -> None:
+    """
+    Remove entries from the LD_PRELOAD environment variable that start
+    with the given prefix.
+
+    Parameters
+    ----------
+    prefix : str
+        The prefix to match entries in LD_PRELOAD.
+
+    """
+    # Get the current LD_PRELOAD value
+    ld_preload = os.environ.get("LD_PRELOAD", "")
+
+    # Split the LD_PRELOAD into a list of entries
+    preload_entries = ld_preload.split(os.pathsep)
+
+    # Remove the entries that start with the given prefix
+    preload_entries = [
+        entry for entry in preload_entries if not entry.startswith(prefix)
+    ]
+
+    # Join the list back into a string and update LD_PRELOAD
+    os.environ["LD_PRELOAD"] = os.pathsep.join(preload_entries)
+
+
+def remove_sox_libmso6_from_ld_preload() -> None:
+    """
+    Remove the sox `libm.so.6` library from the `LD_PRELOAD` environment
+    variable.
+
+    On ubuntu 24.04 the static_sox module does not work with
+    multiprocessing using the spawn method due to a "version
+    GLIBC_2.38 not found" error. This function fixes that by
+    removing the path to the `libm.so.6` library from the `LD_PRELOAD`
+    environment variable.
+
+    """
+    sox_exe = static_sox_run.get_or_fetch_platform_executables_else_raise()
+    remove_from_ld_preload(os.path.join(os.path.dirname(sox_exe), "libm.so.6"))
 
 
 def replace_keys_in_dict(d, old_key_part, new_key_part):
@@ -85,8 +136,10 @@ def load_checkpoint(checkpoint_path, model, optimizer=None, load_opt=1):
     if optimizer and load_opt == 1:
         optimizer.load_state_dict(checkpoint_dict.get("optimizer", {}))
 
-    print(
-        f"Loaded checkpoint '{checkpoint_path}' (epoch {checkpoint_dict['iteration']})",
+    logger.info(
+        "Loaded checkpoint '%s' (epoch %d)",
+        checkpoint_path,
+        checkpoint_dict["iteration"],
     )
     return (
         model,
